refactor: 친구 요청 카운트 타입 Int에서 Long으로 변경

+++ android/app/src/main/java/com/mulkkam/data/repository/FriendsRepositoryImpl.kt
@@ -26,7 +26,7 @@ class FriendsRepositoryImpl(
         )
     }
 
-    override suspend fun getFriendRequestReceivedCount(): MulKkamResult<Int> {
+    override suspend fun getFriendRequestReceivedCount(): MulKkamResult<Long> {
         val result = friendsService.getFriendRequestReceivedCount()
         return result.fold(
             onSuccess = { MulKkamResult(data = it.count) },
